Mac fix double

+++ modules/task_3/tsyplakov_p_monte_karlo/monte_karlo.cpp
@@ -93,6 +93,6 @@ double getParallelMonteKarlo(
     partialResult *= upperLimit[counter] - lowerLimit[counter];
   }
 
-  reductionResult *= (partialResult / (double)amountOfPoint);
+  reductionResult *= (partialResult / amountOfPoint);
   return reductionResult;
 }
